+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -77,6 +77,11 @@ Improvements to clang-tidy
 
   Finds perfect forwarding constructors that can unintentionally hide copy or move constructors.
 
+- New `misc-lambda-function-name
+	<http://clang.llvm.org/extra/clang-tidy/checks/misc-lambda-function-name.html>`_ check
+
+	Finds uses of ``__func__`` or ``__FUNCTION__`` inside lambdas.
+
 - New `modernize-replace-random-shuffle
   <http://clang.llvm.org/extra/clang-tidy/checks/modernize-replace-random-shuffle.html>`_ check